Вынес классы в отдельные файлы и переопределил equals и hashCode для Book, Magazine, Newspaper
