+++ src/main/java/org/example/oop/FinalTaskOOP.java
@@ -5,220 +5,6 @@ import java.util.List;
 import java.util.Scanner;
 import java.util.Objects;
 
-interface Printable {
-    void printDetails();
-}
-
-abstract class Publication implements Printable {
-    private String title;
-    private String author;
-    private int year;
-
-    private static int publicationCount = 0;
-
-    public Publication(String title, String author, int year) {
-        this.title = title;
-        this.author = author;
-        this.year = year;
-        publicationCount++;
-    }
-
-    public static int getPublicationCount() {
-        return publicationCount;
-    }
-
-    public abstract String getType();
-
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public String getAuthor() {
-        return author;
-    }
-
-    public void setAuthor(String author) {
-        this.author = author;
-    }
-
-    public int getYear() {
-        return year;
-    }
-
-    public void setYear(int year) {
-        this.year = year;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        Publication that = (Publication) o;
-        return year == that.year
-                && Objects.equals(title, that.title)
-                && Objects.equals(author, that.author);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(title, author, year);
-    }
-
-    @Override
-    public String toString() {
-        return getType() + " => title=" + title + ", author=" + author + ", year=" + year;
-    }
-}
-
-class Book extends Publication {
-    private String ISBN;
-
-    public Book(String title, String author, int year, String ISBN) {
-        super(title, author, year);
-        this.ISBN = ISBN;
-    }
-
-    public String getISBN() {
-        return ISBN;
-    }
-
-    public void setISBN(String ISBN) {
-        this.ISBN = ISBN;
-    }
-
-    public String getType() {
-        return "Book";
-    }
-
-    @Override
-    public void printDetails() {
-        System.out.println("Тип: " + getType());
-        System.out.println("Название: " + getTitle());
-        System.out.println("Автор: " + getAuthor());
-        System.out.println("Год: " + getYear());
-        System.out.println("ISBN: " + getISBN());
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-        return super.equals(obj);
-    }
-
-    @Override
-    public String toString() {
-        return super.toString() + ", ISBN=" + ISBN;
-    }
-}
-
-class Magazine extends Publication {
-    private int IssueNumber;
-
-    public Magazine(String title, String author, int year, int issueNumber) {
-        super(title, author, year);
-        this.IssueNumber = issueNumber;
-    }
-
-    public int getIssueNumber() {
-        return IssueNumber;
-    }
-
-    public void setIssueNumber(int issueNumber) {
-        IssueNumber = issueNumber;
-    }
-
-    public String getType() {
-        return "Журнал";
-    }
-
-    @Override
-    public void printDetails() {
-        System.out.println("Тип: " + getType());
-        System.out.println("Название: " + getTitle());
-        System.out.println("Автор: " + getAuthor());
-        System.out.println("Год: " + getYear());
-        System.out.println("Номер выпуска: " + getIssueNumber());
-    }
-
-    @Override
-    public String toString() {
-        return super.toString() + ", IssueNumber=" + IssueNumber;
-    }
-}
-
-class Newspaper extends Publication {
-    private String publicationDay;
-
-    public Newspaper(String title, String author, int year, String publicationDay) {
-        super(title, author, year);
-        this.publicationDay = publicationDay;
-    }
-
-    public String getPublicationDay() {
-        return publicationDay;
-    }
-
-    public void setPublicationDay(String publicationDay) {
-        this.publicationDay = publicationDay;
-    }
-
-    public String getType() {
-        return "Газета";
-    }
-
-    @Override
-    public void printDetails() {
-        System.out.println("Тип: " + getType());
-        System.out.println("Название: " + getTitle());
-        System.out.println("Автор: " + getAuthor());
-        System.out.println("Год: " + getYear());
-        System.out.println("День издания: " + getPublicationDay());
-    }
-
-    @Override
-    public String toString() {
-        return super.toString() + ", publicationDay=" + publicationDay;
-    }
-}
-
-class Library {
-    private List<Publication> publications;
-
-    public Library() {
-        this.publications = new ArrayList<Publication>();
-    }
-
-    public void addPublication(Publication pub) {
-        publications.add(pub);
-    }
-
-    public void listPublications() {
-        if (publications.isEmpty()) {
-            System.out.println("Список публикаций пуст!");
-        } else {
-            for (int i = 0; i < publications.size(); i++) {
-                System.out.println((i + 1) + ") " + publications.get(i));
-            }
-        }
-    }
-
-    public void searchByAuthor(String author) {
-        if (author == null) {
-            System.out.println("Ошибка! Автор не указан");
-        } else {
-            for (Publication pub : publications) {
-                if (pub.getAuthor().equals(author)) {
-                    System.out.println(pub);
-                }
-            }
-        }
-
-    }
-}
-
 public class FinalTaskOOP {
 
     public static void main(String[] args) {
