+++ src/main/java/org/example/oop/Printable.java
@@ -1,0 +1,5 @@
+package org.example.oop;
+
+interface Printable {
+    void printDetails();
+}
