+++ src/main/java/org/example/oop/Publication.java
@@ -1,0 +1,68 @@
+package org.example.oop;
+
+import java.util.Objects;
+
+abstract class Publication implements Printable {
+    private String title;
+    private String author;
+    private int year;
+
+    private static int publicationCount = 0;
+
+    public Publication(String title, String author, int year) {
+        this.title = title;
+        this.author = author;
+        this.year = year;
+        publicationCount++;
+    }
+
+    public static int getPublicationCount() {
+        return publicationCount;
+    }
+
+    public abstract String getType();
+
+    public String getTitle() {
+        return title;
+    }
+
+    public void setTitle(String title) {
+        this.title = title;
+    }
+
+    public String getAuthor() {
+        return author;
+    }
+
+    public void setAuthor(String author) {
+        this.author = author;
+    }
+
+    public int getYear() {
+        return year;
+    }
+
+    public void setYear(int year) {
+        this.year = year;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        Publication that = (Publication) o;
+        return year == that.year
+                && Objects.equals(title, that.title)
+                && Objects.equals(author, that.author);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(title, author, year);
+    }
+
+    @Override
+    public String toString() {
+        return getType() + " => title=" + title + ", author=" + author + ", year=" + year;
+    }
+}
